Refactor ball by pulling out enemyPosition into its own class
