+++ lib/ball.js
@@ -1,20 +1,7 @@
-function setPosition() {
-  var newPosition = null;
-  var newX        = null;
-
-  var chance = Math.random();
-  if (chance > 0.5) {
-    newPosition = 1;
-    newX        = 410;
-  } else {
-    newPosition = 2;
-    newX        = 490;
-  }
-  return { position: newPosition, x: newX };
-}
+var enemyPosition = require('./enemy_position');
 
 function Ball(params) {
-  this.newPosition      = setPosition();
+  this.newPosition      = enemyPosition();
   this.currentPosition  = this.newPosition.position;
   this.nextPosition     = this.newPosition.position;
   this.board            = params['board'];
